tests/wpt/web-platform-tests/html/browsers/offline/application-cache-api/api_status_updateready-manual.html

<!DOCTYPE HTML>
<html manifest="../resources/manifest/clock.manifest">
  <head>
    <title>Offline Application Cache - API_status_UPDATEREADY</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <ol>
      <li>Modify the commented part in the manifest file (manifest/clock.manifest) on the server.</li>
      <li>Refresh the page.</li>
    </ol>

    <div id="log"></div>

    <script>
      var t = async_test("updateready status test"),
          cache = window.applicationCache;

      cache.onupdateready = t.step_func_done(function() {
        assert_equals(cache.status, cache.UPDATEREADY, "cache.status should equals cache.UPDATEREADY");
      });
    </script>
  </body>
</html>
